Refactor libusb driver
